requirements.txt
playwright
python-dotenv
azure-storage-file-datalake
